Add loopback, sourcesink, ffs support

+++ setup_usb_gadget.sh
@@ -7,6 +7,28 @@ APP=~/aio_simple
 idVendor=0xfffe
 idProduct=0xa4a4
 
+function usage() {
+    echo "Usage: $0 [-l|-s|-f]"
+    echo "  Enable one or more usb gadget functions:"
+    echo "  -l Loopback"
+    echo "  -s SourceSink"
+    echo "  -f FFS"
+}
+
+function parse_args() {
+    [[ $# == 0 ]] && usage && exit
+    for arg in "$@"
+    do
+        case "$arg" in
+            "-l") lb=1 ;;
+            "-s") ss=1 ;;
+            "-f") ffs=1 ;;
+            *) usage;
+               exit ;;
+        esac
+    done
+}
+
 function ffs_app() {
     # Caution: Starting and stopping the ffs app is very application specific.
     # It is very likely that the start and stop cases will need to be adjusted
@@ -15,7 +37,7 @@ function ffs_app() {
         stop) killall -s SIGTERM $(basename $APP) ;;
         *) echo "ffs_app(): unknown pattern" ;;
     esac
-    sleep 0.01
+    sleep 0.1
 }
 
 function init_usb_gadget() {
@@ -27,15 +49,27 @@ function init_usb_gadget() {
     mkdir configs/c.1
     mkdir configs/c.1/strings/0x409
     echo 2 > configs/c.1/MaxPower
-    mkdir functions/ffs.usb0
-    ln -s functions/ffs.usb0 configs/c.1/
 
-    # create functionfs
-    mkdir -p $FFS_DIR
-    mount usb0 $FFS_DIR -t functionfs
+    if [[ $ffs ]]; then
+        # link ffs
+        mkdir functions/ffs.usb0
+        ln -s functions/ffs.usb0 configs/c.1/
+        mkdir -p $FFS_DIR
+        mount usb0 $FFS_DIR -t functionfs
+        ffs_app start
+    fi
 
-    # start functionfs application
-    ffs_app start
+    if [[ $lb ]]; then
+        # link loopback
+        mkdir functions/Loopback.usb0
+        ln -s functions/Loopback.usb0 configs/c.1/
+    fi
+    
+    if [[ $ss ]]; then
+        # link sourcesink
+        mkdir functions/SourceSink.usb0
+        ln -s functions/SourceSink.usb0 configs/c.1/
+    fi
 
     # start usb gadget
     echo $UDC > $G1/UDC
@@ -45,23 +79,39 @@ function deinit_usb_gadget() {
     # stop usb gadget
     echo "" > $G1/UDC
 
-    # stop functionfs application
-    ffs_app stop
-
     # clean up
-    umount $FFS_DIR
-    rmdir $FFS_DIR
     cd $G1
-    rm configs/c.1/ffs.usb0
+    
+    if [[ -d "configs/c.1/ffs.usb0" ]]; then
+        #unlink ffs
+        ffs_app stop
+        umount $FFS_DIR
+        rmdir $FFS_DIR
+        rm configs/c.1/ffs.usb0
+        rmdir functions/ffs.usb0
+    fi
+
+    if [[ -d "configs/c.1/Loopback.usb0" ]]; then
+        #unlink loopback
+        rm configs/c.1/Loopback.usb0
+        rmdir functions/Loopback.usb0
+    fi
+    
+    if [[ -d "configs/c.1/SourceSink.usb0" ]]; then
+        #unlink sourcesink
+        rm configs/c.1/SourceSink.usb0
+        rmdir functions/SourceSink.usb0
+    fi
+    
     rmdir configs/c.1/strings/0x409
     rmdir configs/c.1
-    rmdir functions/ffs.usb0
     rmdir strings/0x409
     cd ~
     rmdir $G1
 }
 
 if [[ ! -d $G1 ]]; then
+    parse_args $@
     init_usb_gadget
 else
     read -r -p "Remove gadget? [yN]" a
